+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `claude-sonnet-4.5` for github-copilot provider.
+- Add `prompt-received` metric.
 
 ## 0.62.0